Add speedier replacement for BasicMemoryCache
Add BetterBasicMemoryCache to replace BasicMemoryCache which has been renamed to OldBasicMemoryCache. BetterBasicMemoryCache--yes, I know the name is stupid--uses ConcurrentLinkedHashMap and doesn't require an expensive Arrays.sort call when evicting things from the cache. This was really bad when it happened on the UI thread and also caused all cache read operations to block while it was happening